Consolidated my styled component files into one. Modified my component to use our new server routes.
