--- client/src/Ratings-Reviews/RR-styled-components/AverageRatingStyle.jsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const AverageRating = styled.canvas`
-align-self: center;
-flex-flow: column wrap;
-border-style: solid;
-border-color: green;
-`;
-
-export default AverageRating;
