--- client/src/Ratings-Reviews/RR-styled-components/BarChartStyle.jsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const BarChartStyle = styled.div`
-align-self: center;
-width:30%;
-flex-flow: column wrap;
-border-style: solid;
-border-color: green;
-`;
-
-export default BarChartStyle;
