--- client/src/Ratings-Reviews/RR-styled-components/PercentageOfRecsStyle.jsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const Percentage = styled.div`
-align-self: flex-start;
-flex-flow: column wrap;
-border-style: solid;
-border-color: gray;
-`;
-
-export default Percentage;
