--- client/src/Ratings-Reviews/RR-styled-components/ProductBreakdownContainer.jsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const ProductBreakdownContainer = styled.div`
-align-self: flex-start;
-flex-flow: column wrap;
-border-style: solid;
-border-color: pink;
-`;
-
-export default ProductBreakdownContainer;
