--- client/src/Ratings-Reviews/RR-styled-components/RatingSummary.jsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const SummaryLeft = styled.div`
-align-self: flex-start;
-width: 30%;
-flex-flow: column wrap;
-border-style: solid;
-border-color: red;
-`;
-
-export default SummaryLeft;
