--- client/src/Ratings-Reviews/RR-styled-components/RatingsBreakdown.jsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const RatingsBreakdown = styled.div`
-align-self: flex-start;
-flex-flow: row wrap;
-border-style: solid;
-border-color: red;
-`;
-
-export default RatingsBreakdown;
