--- client/src/Ratings-Reviews/RR-styled-components/ReviewEntryStyle.jsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const ReviewEntryStyle = styled.div`
-align-self: flex-start;
-flex-flow: column wrap;
-border-style: solid;
-`;
-
-export default ReviewEntryStyle;
