--- client/src/Ratings-Reviews/RR-styled-components/ReviewListContainer.jsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const ReviewListContainer = styled.div`
-align-self: flex-start;
-width: 66%;
-flex-flow: column wrap;
-border-style: solid;
-border-color: cyan;
-`;
-
-export default ReviewListContainer;
