+++ client/src/Ratings-Reviews/RRcomponents/RatingSummary/PercentageOfRecs.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Percentage from '../../RR-styled-components/PercentageOfRecsStyle';
+import { Percentage } from '../../RR-styled-components/RRsectionContainerStyle';
 
 export default function PercentageOfRecs() {
   return (
